pyproject.toml
[project]
name = "gemini-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai==1.72.0",
    "python-dotenv==1.2.2",
]
